pyproject.toml
[tool.poetry]
name = "project.nosync"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
requests = "^2.28.1"
keras = "^2.10.0"
Pillow = "^9.2.0"
pandas = "^1.4.4"
scikit-learn = "^1.1.2"
matplotlib = "^3.6.0"
tensorflow = "^2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
